backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.27
asyncpg==0.29.0
pgvector==0.2.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.3.2
pypdf2==3.0.1
python-docx==1.1.0
alembic==1.13.1
